+++ packages/block-library/src/pullquote/edit.js
@@ -18,7 +18,7 @@ import {
 	InspectorControls,
 	withColors,
 	PanelColorSettings,
-} from '@wordpress/editor';
+} from '@wordpress/block-editor';
 
 export const SOLID_COLOR_STYLE_NAME = 'solid-color';
 export const SOLID_COLOR_CLASS = `is-style-${ SOLID_COLOR_STYLE_NAME }`;