+++ packages/e2e-tests/specs/blocks/list.test.js
@@ -3,6 +3,7 @@
  */
 import {
 	clickBlockAppender,
+	clickBlockToolbarButton,
 	getEditedPostContent,
 	createNewPost,
 	pressKeyTimes,
@@ -124,11 +125,7 @@ describe( 'List', () => {
 		await insertBlock( 'List' );
 		await page.keyboard.type( 'one' );
 		await page.keyboard.press( 'Enter' );
-		// Pointer device is needed. Shift+Tab won't focus the toolbar.
-		// To do: fix so Shift+Tab works.
-		await page.mouse.move( 200, 300, { steps: 10 } );
-		await page.mouse.move( 250, 350, { steps: 10 } );
-		await page.click( 'button[aria-label="Indent list item"]' );
+		await clickBlockToolbarButton( 'Indent list item' );
 		await page.keyboard.type( 'two' );
 		await transformBlockTo( 'Paragraph' );
 
@@ -198,11 +195,7 @@ describe( 'List', () => {
 		await insertBlock( 'List' );
 		await page.keyboard.type( 'one' );
 		await page.keyboard.press( 'Enter' );
-		// Pointer device is needed. Shift+Tab won't focus the toolbar.
-		// To do: fix so Shift+Tab works.
-		await page.mouse.move( 200, 300, { steps: 10 } );
-		await page.mouse.move( 250, 350, { steps: 10 } );
-		await page.click( 'button[aria-label="Indent list item"]' );
+		await clickBlockToolbarButton( 'Indent list item' );
 		await page.keyboard.type( 'two' );
 		await page.keyboard.press( 'Enter' );
 		await page.keyboard.type( 'three' );
@@ -233,12 +226,7 @@ describe( 'List', () => {
 		await pressKeyWithModifier( 'primary', 'm' );
 		await page.keyboard.type( '1' );
 
-		// Pointer device is needed. Shift+Tab won't focus the toolbar.
-		// To do: fix so Shift+Tab works.
-		await page.mouse.move( 200, 300, { steps: 10 } );
-		await page.mouse.move( 250, 350, { steps: 10 } );
-
-		await page.click( 'button[aria-label="Convert to ordered list"]' );
+		await clickBlockToolbarButton( 'Convert to ordered list' );
 
 		expect( await getEditedPostContent() ).toMatchSnapshot();
 	} );