+++ packages/components/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 7.1.0 (Unreleased)
+## 7.1.0 (2019-03-06)
 
 ### New Features